feat: Migrate test runner in channel/workflows from jest to vitest
