+++ services/channel/workflows/src/stations/Channels/ProgramManagement/CuePointSchedule/CuePointSchedule.tsx
@@ -1,6 +1,5 @@
 // TODO: Remove these disables
- 
- 
+
 import {
   ActionData,
   formatSecondsToTimestamp,
