+++ services/channel/workflows/tsconfig.json
@@ -7,10 +7,7 @@
     "sourceMap": true,
     "outDir": "./dist",
     "skipLibCheck": true,
-    "lib": [
-      "dom",
-      "es2018"
-    ],
+    "lib": ["dom", "es2018"],
     "moduleResolution": "bundler",
     "module": "esnext",
     "jsx": "react",
@@ -19,10 +16,6 @@
     "esModuleInterop": true,
     "resolveJsonModule": true
   },
-  "include": [
-    "./src"
-  ],
-  "exclude": [
-    "node_modules"
-  ]
+  "include": ["./src"],
+  "exclude": ["node_modules"]
 }
